refactor(core): share DataProtection between Bff and Api
Move DataProtectionOptions to affolterNET.Web.Core.Configuration and
introduce IServiceCollection.AddAzureBlobDataProtection so both Web.Bff
and Web.Api can persist DP keys to Azure Blob Storage with the same
affolterNET:Web:DataProtection config contract. Wires it into ApiAppOptions
and AddApiServices so JWT-bearer APIs can opt in via the same env vars.
