+++ affolterNET.Web.Bff/Extensions/ServiceCollectionExtensions.cs
@@ -3,12 +3,9 @@ using affolterNET.Web.Bff.Configuration;
 using affolterNET.Web.Bff.Middleware;
 using affolterNET.Web.Bff.Options;
 using affolterNET.Web.Bff.Services;
-using Azure.Identity;
-using Azure.Storage.Blobs;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authentication.OpenIdConnect;
-using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Configuration;
@@ -56,7 +53,7 @@ public static class ServiceCollectionExtensions
         services.AddBffAuthenticationInternal(bffOptions);
 
         // Add Data Protection key persistence
-        services.AddDataProtectionInternal(bffOptions);
+        services.AddAzureBlobDataProtection(bffOptions.DataProtection);
 
         // Add BFF supporting services
         services.AddAntiforgeryServicesInternal(bffOptions.AntiForgery);
@@ -328,44 +325,6 @@ public static class ServiceCollectionExtensions
                 : Microsoft.AspNetCore.Http.CookieSecurePolicy.SameAsRequest;
             options.Cookie.Path = bffAntiforgeryOptions.CookiePath;
         });
-
-        return services;
-    }
-
-    /// <summary>
-    /// Adds Data Protection key persistence to Azure Blob Storage.
-    /// When disabled (default), ASP.NET Core uses ephemeral in-memory keys.
-    /// </summary>
-    private static IServiceCollection AddDataProtectionInternal(
-        this IServiceCollection services, BffAppOptions bffOptions)
-    {
-        if (!bffOptions.DataProtection.Enabled)
-        {
-            return services;
-        }
-
-        var dp = bffOptions.DataProtection;
-        var dpBuilder = services.AddDataProtection();
-
-        if (!string.IsNullOrWhiteSpace(dp.ApplicationName))
-        {
-            dpBuilder.SetApplicationName(dp.ApplicationName);
-        }
-
-        if (!string.IsNullOrWhiteSpace(dp.StorageAccountName))
-        {
-            // Managed Identity auth
-            var blobUri = new Uri(
-                $"https://{dp.StorageAccountName}.blob.core.windows.net/{dp.ContainerName}/{dp.BlobName}");
-            var credential = new ManagedIdentityCredential(dp.ManagedIdentityClientId);
-            dpBuilder.PersistKeysToAzureBlobStorage(blobUri, credential);
-        }
-        else if (!string.IsNullOrWhiteSpace(dp.ConnectionString))
-        {
-            // Connection string auth (local docker, testing)
-            var blobClient = new BlobClient(dp.ConnectionString, dp.ContainerName, dp.BlobName);
-            dpBuilder.PersistKeysToAzureBlobStorage(blobClient);
-        }
 
         return services;
     }
